#6 Add Navigation Agent to Hero And Update Goon and Hero to CharacterBody2D
Added navigation agent to hero but it wasn't avoiding. So I added some collision shapes (ended up being unnecessary) and figured I'd make them CharacterBody2D because that made sense for what they would be.

So, right now "Avoid" is on for the Goon AND hero. However, the Goons don't actually calculate the safe velocity, so they still just go straight. It might be screwing with the pathing of the Hero. Will have to do some masking stuff so that the Hero avoids Goons but the Goons don't avoid HEro.