#6 Added Goon Chase Hero NOT QUITE WORKING
For some reason the Goons won't actually finish the hero off. I thought at first it was their avoidant behavior, but now I'm thinking it's because the "safe velocity" is taking the hero OUTSIDE of the navigation region. So the little guys can't run out there they won't path outside of the region but the "safe velocity" is letting him out. So, I'm gonna try making a tile map so that I can SEE where the navigation plane is.