#6 Added Desire Back to Center And Misc fixes
Fixed a few things (like the current collide circle radius being just baarreeelly biger than the hurtbox at one point so they could get stuck on each other).

Also made it to where to Hero has a desire to be in teh center of the map to help him not get stuck in corners so badly.